+++ wallet-dist/assets/popup.html-C8PKVe3_.js
@@ -1,0 +1,1 @@
+import"./popup-BajrIw1T.js";